Add missing cases for expression diagnostics
